extract the core data validation from the script
